ERR: type conversion warnings.

+++ Testing/Code/IO/testMetaBlob.cxx
@@ -38,7 +38,9 @@ int testMetaBlob(int , char * [])
   for(i=0;i<10;i++)
   {
     pnt = new BlobPnt(3);
-    pnt->m_X[0]=(float)0.2;pnt->m_X[1]=i;pnt->m_X[2]=i;
+    pnt->m_X[0]=(float)0.2;
+    pnt->m_X[1]=static_cast<float>(i);
+    pnt->m_X[2]=static_cast<float>(i);
     blob.GetPoints().push_back(pnt);
   }
